backend/requirements.txt
alembic==1.13.1
asyncpg==0.29.0
bcrypt==4.1.2
fastapi==0.110.0
greenlet==3.0.3
httpx==0.27.0
passlib==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.3
SQLAlchemy==2.0.28
uvicorn==0.29.0
pytest==8.1.1
pytest-asyncio==0.23.5
freezegun==1.4.0
email-validator==2.1.1
